Corrected: Use long pk in set! Issue: DCMEE-270

+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-web/src/java/org/dcm4chex/archive/web/maverick/FolderUtil.java
@@ -93,7 +93,7 @@ public class FolderUtil {
 		if ( stickyInstances.size() > 0 ) {
 			Set set = new HashSet();
 			for ( Iterator iter = stickyInstances.iterator() ; iter.hasNext() ; ) {
-				set.add( Integer.valueOf(iter.next().toString()));
+				set.add( Long.valueOf(iter.next().toString()));
 			}
 			try {
 				Collection col = lookupContentManager().getSOPInstanceRefMacros(set);
